backend/pyproject.toml
[tool.poetry]
name = "shortdiary"
version = "1.0.0"
description = ""
authors = ["Lukas Martini <hello@lutoma.org>"]
license = "Proprietary"

[tool.poetry.dependencies]
python = "3.9.2"
Django = "^3.1"
Pillow = "^8.1.2"
django-formtools = "^2.0"
django-gravatar2 = "^1.4.4"
djangorestframework = "^3.4.7"
gunicorn = "^19.1.1"
psycopg2-binary = "^2.8.6"
python-memcached = "^1.59"
pytz = "^2021.1"
phonenumbers = "^8.12.20"
guess_language-spirit = "^0.5.3"
django-two-factor-auth = "^1.13"
twilio = "^6.55.0"
django-q = "^1.3.5"
redis = "^3.5.3"
pystmark = "^0.5.0"
django-otp-yubikey = "^1.0.0"
pycrypto = "^2.6.1"
django-redis = "^4.12.1"
django-rest-auth = "^0.9.5"
django-cors-headers = "^3.7.0"
django-storages = "^1.11.1"
boto3 = "^1.17.45"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
